refactor: merge child-school-overview components into one
to unite the separate components for a school's students and a student's schools
