--- src/app/child-dev-project/children/previous-schools/previous-schools.component.html
@@ -1,17 +1,0 @@
-<app-pill class="hint-badge" *ngIf="!hasCurrentlyActiveEntry">
-  Currently there is no active entry. To add a new entry, click on the
-  <fa-icon
-    class="pointer color-accent"
-    aria-label="add"
-    icon="plus-circle"
-    (click)="entitySubrecord.create()"></fa-icon>
-  button.
-</app-pill>
-
-<app-entity-subrecord
-  #entitySubrecord
-  [records]="records"
-  [columns]="columns"
-  [newRecordFactory]="generateNewRecordFactory()"
->
-</app-entity-subrecord>
