+++ src/app/child-dev-project/schools/child-school-overview/child-school-overview.component.html
@@ -1,0 +1,33 @@
+<app-pill class="hint-badge" *ngIf="!hasCurrentlyActiveEntry && mode === 'child'">
+  Currently there is no active entry. To add a new entry, click on the
+  <fa-icon
+    class="pointer color-accent"
+    aria-label="add"
+    icon="plus-circle"
+    (click)="entitySubrecord.create()"></fa-icon>
+  button.
+</app-pill>
+
+<app-entity-subrecord
+  #entitySubrecord
+  [records]="displayedRecords"
+  [isLoading]="isLoading"
+  [columns]="columns"
+  [newRecordFactory]="generateNewRecordFactory()"
+  [getBackgroundColor]="backgroundColorFn"
+>
+</app-entity-subrecord>
+
+<div class="margin-top-regular">
+  <mat-slide-toggle
+    [(ngModel)]="showInactive"
+    (change)="prepareDisplayedData()"
+    i18n-matTooltip="Show ended tooltip|Tooltip that will appear when hovered over the show-ended button"
+    matTooltip="Show entries where the end date is in the past"
+    i18n="slider|also show entries that are inactive"
+  >
+    Include inactive past records
+  </mat-slide-toggle>
+</div>
+
+
