+++ src/app/child-dev-project/schools/child-school-overview/child-school-overview.component.scss
