--- src/app/child-dev-project/schools/children-overview/children-overview.component.html
@@ -1,20 +1,0 @@
-<app-entity-subrecord
-  [records]="displayedRecords"
-  [columns]="columns"
-  [newRecordFactory]="generateNewRecordFactory()"
-  [getBackgroundColor]="backgroundColorFn"
-></app-entity-subrecord>
-
-<div style="margin-top: 16px">
-  <mat-slide-toggle
-    [(ngModel)]="showInactive"
-    (change)="this.prepareDisplayedData()"
-    i18n-matTooltip="Show ended tooltip|Tooltip that will appear when hovered over the show-ended button"
-    matTooltip="Show entries where the end date is in the past"
-    i18n="slider|also show entries that are inactive"
-  >
-    Include ended ones
-  </mat-slide-toggle>
-</div>
-
-
